update combo text show immediatly when cluster start happen
